refactor: location of where types are saved
