+++ workspaces/tech-insights/plugins/tech-insights-common/src/types/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 The Backstage Authors
+ * Copyright 2025 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,12 +13,5 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-import * as anything from './';
-
-describe('tech-insights-common', () => {
-  // Types only
-  it('should exist', () => {
-    expect(anything).toBeTruthy();
-  });
-});
+export * from './checks';
+export * from './facts';
